requirements.txt
Django==3.1.4
django-extensions==3.1.0
ipython==7.19.0
django-debug-toolbar==3.2

